# GeLoRa/backend/docs/swagger/swagger.yaml
definitions:
  adaptersinboundhttpdto.CommonErrorResponse:
    properties:
      error:
        type: string
    type: object
  adaptersinboundhttpdto.CommonPaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  adaptersinboundhttpdto.GetNodeInfoByIdResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Primary entry node.
        type: string
      id:
        example: 1
        type: integer
      is_validated:
        example: true
        type: boolean
      mid:
        example: node-abc123
        type: string
      name:
        example: Main Gateway
        type: string
      updated_at:
        example: "2024-06-01T00:00:00Z"
        type: string
      validated_by:
        example: 2
        type: integer
    type: object
  adaptersinboundhttpdto.GetNodesListData:
    properties:
      id:
        example: 1
        type: integer
      is_validated:
        example: true
        type: boolean
      mid:
        example: node-abc123
        type: string
      name:
        example: Main Gateway
        type: string
    type: object
  adaptersinboundhttpdto.GetNodesListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/adaptersinboundhttpdto.GetNodesListData'
        type: array
      pagination:
        $ref: '#/definitions/adaptersinboundhttpdto.CommonPaginationMeta'
    type: object
...
  adaptersinboundhttpdto.PatchNodeInfoByIdRequest:
    properties:
      description:
        example: Primary entry node.
        type: string
      name:
        example: Main Gateway
        type: string
    type: object
...
info:
  contact: {}
  description: The core API documentation for GeLoRa backend endpoints.
  title: GeLoRa API
  version: "1.0"
paths:
...
  /nodes:
    get:
      description: Return a paginated list of nodes with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Cursor ID for pagination
        in: query
        name: cursor_id
        type: integer
      - description: Search by name or MID
        in: query
        name: search
        type: string
      - description: Filter by validation status
        in: query
        name: is_validated
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.GetNodesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: List nodes
      tags:
      - Nodes
  /nodes/{id}:
    delete:
      description: Permanently remove a node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete node by ID
      tags:
      - Nodes
    get:
      description: Return the full details of a specific node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.GetNodeInfoByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Get node by ID
      tags:
      - Nodes
    patch:
      consumes:
      - application/json
      description: Update the name and/or description of a specific node
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adaptersinboundhttpdto.PatchNodeInfoByIdRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Update node by ID
      tags:
      - Nodes
  /nodes/{id}/validate:
    put:
      description: Mark a node as validated by the currently authenticated user
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/adaptersinboundhttpdto.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate node by ID
      tags:
      - Nodes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"